requirements.txt
Flask==2.3.3
tensorflow-cpu==2.18.0
Pillow==10.0.1
opencv-python==4.8.1.78
matplotlib==3.7.2
mysql-connector-python==8.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
numpy==1.24.3